+++ R/RcppExports.R
@@ -1059,7 +1059,9 @@ growth_day <- function(x, date, meteovec, latitude, elevation, slope, aspect, ru
 #' 
 #' #Switch to 'Cochard' transpiration mode
 #' control <- defaultControl("Cochard")
-#' control$equalLeafStemVC = TRUE # Makes leaf xylem vulnerability equal to stem xylem vulnerability
+#' 
+#' #Makes leaf xylem vulnerability equal to stem xylem vulnerability
+#' control$equalLeafStemVC = TRUE 
 #' 
 #' #Initialize vegetation input
 #' x3 <- forest2growthInput(exampleforestMED,examplesoil, SpParamsMED, control)
@@ -3250,7 +3252,7 @@ spwb_day <- function(x, date, meteovec, latitude, elevation, slope, aspect, runo
 #' 
 #' @seealso 
 #' \code{\link{spwbInput}}, \code{\link{spwb_day}}, \code{\link{plot.spwb}}, 
-#' \code{\link{extractDaily}}, \code{\link{summary.spwb}},  \code{\link{forest}}, \code{\link{aspwb}}
+#' \code{\link{extract}}, \code{\link{summary.spwb}},  \code{\link{forest}}, \code{\link{aspwb}}
 #' 
 #' @examples
 #' #Load example daily meteorological data
@@ -3286,7 +3288,9 @@ spwb_day <- function(x, date, meteovec, latitude, elevation, slope, aspect, runo
 #' 
 #' #Switch to 'Cochard' transpiration mode
 #' control <- defaultControl("Cochard")
-#' control$equalLeafStemVC = TRUE # Makes leaf xylem vulnerability equal to stem xylem vulnerability
+#' 
+#' # Makes leaf xylem vulnerability equal to stem xylem vulnerability
+#' control$equalLeafStemVC = TRUE 
 #' 
 #' #Initialize input
 #' x3 <- forest2spwbInput(exampleforestMED,examplesoil, SpParamsMED, control)